Don't alter filesystem rw state, it's pointless
SD cards run maintenance tasks on the flash at arbitrary times for
wear leveling and to deal with read disturb[1]. This is one of the
real reasons the Pi has issues with SD card corruption, sudden power
loss can randomly interrupt that maintenance, and undervoltage can
cause cheap SD card controllers to malfunction.

You can't prevent any of that by making the filesystem read-only,
the card has no idea whether the filesystem is mounted that way or
not and will carry out maintenance tasks whenever it wants.

The only way to prevent that from causing filesystem corruption is to
use SD cards that can handle sudden power loss, like industrial cards
that are generally designed for it, unlike commercial SD cards.

A better solution to issues that do affect the filesystem layer would
be to significantly shorten the journal flush time.

[1] https://en.wikipedia.org/wiki/Flash_memory#Read_disturb